requirements.txt
sentence-transformers
faiss-cpu
fastapi
uvicorn
pdfminer.six
python-docx
pillow
pytesseract
nltk